+++ sockeye/train.py
@@ -291,7 +291,7 @@ def create_data_iters_and_vocabs(args: argparse.Namespace,
             batch_size=args.batch_size,
             batch_by_words=batch_by_words,
             batch_num_devices=batch_num_devices,
-            batch_sentences_multiple_of=args.round_batch_sizes_to_multiple_of)
+            batch_sentences_multiple_of=args.batch_sentences_multiple_of)
 
         check_condition(all([combine in [C.SOURCE_FACTORS_COMBINE_SUM, C.SOURCE_FACTORS_COMBINE_AVERAGE]
                              for combine in args.source_factors_combine])
@@ -379,7 +379,7 @@ def create_data_iters_and_vocabs(args: argparse.Namespace,
             bucketing=not args.no_bucketing,
             bucket_width=args.bucket_width,
             bucket_scaling=not args.no_bucket_scaling,
-            batch_sentences_multiple_of=args.round_batch_sizes_to_multiple_of)
+            batch_sentences_multiple_of=args.batch_sentences_multiple_of)
 
         data_info_fname = os.path.join(output_folder, C.DATA_INFO)
         logger.info("Writing data config to '%s'", data_info_fname)