+++ CHANGELOG.md
@@ -11,6 +11,14 @@ Note that Sockeye has checks in place to not translate with an old model that wa
 
 Each version section may have have subsections for: _Added_, _Changed_, _Removed_, _Deprecated_, and _Fixed_.
 
+## [2.1.5]
+
+### Changed
+
+- Batch sizes are now per device (ex: `--batch-size 2048 --device-ids -4` gives an effective batch size of 8192).
+- Word-based batch sizes are now a max and include padding tokens in the count.  This makes memory usage more predictable.
+- Renamed `--round-batch-sizes-to-multiple-of` to `--batch-sentences-multiple-of` for clarity.  Sizes are always rounded down.
+
 ## [2.1.4]
 
 ### Added